feat: :rocket: point component
The what
Two components, user.points.component, which receives prop changes in points and sends the value to point.component, which displays the current points.

The why
Take in and display point transactions.

Uncertainties
Lots. My thoughts are to have user.points.component act as a transaction parser, reading changes in props, the corresponding transactionID (commented out in the service-worker as im not sure how to implement it yet/want your thoughts), and perhaps store/display transaction history and timestamps in an array. I need to update the nomenclature as well as it is messy/unintuitive.