feat: :rocket: fix: :bug:
The what:
Removed user.points.component. Added circular progress bar, with a cap of 4 points, increasing in 25% increments

The why:
user.points.component was unnecessary, progress bar to visualize incrementing points.

Uncertainties:
Need to add a percentage to completion value, used yarn to add this component, not sure if thats an issue if you use npm