feat: :rocket: Service worker coffee notification
The what:
Added a coffeeCount variable to service-worker state, and isReturn() function. isReturn() gets called in createNotification(), and it increments coffeeCount, resets it to 0 if the value is 4 and returns that you have enough for a coffee, else it returns 4 - coffeeCount "points to get a coffee", which gets used in the body of the notification. Also reduced notification timer to each 20 seconds.

The Why:
To make notifications more realistic, and accurately inform the user how many points they are from a coffee/when they have enough for a coffee. Notification timer reduced for faster debugging/demo.

Uncertanties:
Unsure if points to get a coffee should be tracked in the service-worker, or on the app side, and create a notification when user has points = multiple of 4.