fix: :bug: Updated notification messages
The What:
Removed the counter in service-worker, Moved isReward() function to app.js, changed it to depend on modulo of user points instead of counter resetting to 0,
Service-worker notification body parameter is now dependent on this.props.notificationText.

The Why:
Very unintuitive to have a separate counter in the service-worker to keep track of points as well as in App.js.

Uncertainties:
Not sure if it would be better to have isReward() be a part of app.js or within the service-worker.