fix: 🐛Notification pass as function
The What:
Updated serviceworker call to pass isReward, instead of isReward()

The Why:
Executing the function instead of passing it as a parameter can lead to issues in the future.

Uncertaintes:
Might be better to have notificationText() executed and stored in a variable in service worker, as opposed to jsut calling it in the body definition.